Collated CrewMember tests and fixed bugs
Added a new test set called CrewMemberParameterizedTest which creates a
list of crew members from which to test from instead of creating
individual test files for each crew member type to avoid repeating code.
Also takes input from a text file to run all the tests quickly.
Added a setScanner method in Ship to assist this.
The old tests are still available and should work just in case.
Changed some of the assertTrue statements to assertEquals because i'm
dumb and it works fine.
Made actions use nutrition as specified in the design document.
Wrote some comments in some of the longer functions in CrewMember.